wip: list of students redesign
